Fixes big text command crashing in private channels

+++ src/main/java/com/darkzek/ChickenBot/Commands/BigText.java
@@ -5,7 +5,9 @@ import com.darkzek.ChickenBot.Enums.TriggerType;
 import com.darkzek.ChickenBot.Events.CommandRecievedEvent;
 import com.darkzek.ChickenBot.Settings;
 import com.darkzek.ChickenBot.Trigger;
+import net.dv8tion.jda.core.entities.ChannelType;
 import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
+import net.dv8tion.jda.core.exceptions.PermissionException;
 
 import javax.imageio.ImageIO;
 import java.awt.*;
@@ -28,68 +30,65 @@ public class BigText extends Command {
     @Override
     public void MessageRecieved(CommandRecievedEvent event) {
 
-        event.getMessage().delete().queue();
+        if (event.getChannelType() == ChannelType.TEXT) {
+            try {
+                event.getMessage().delete().queue();
+            } catch (PermissionException e) { }
+        }
 
-        new Thread(new Runnable() {
+        new Thread(() -> {
 
-            @Override
-            public void run() {
+            //Get between all the spaces, to remove the command bit
 
-                String message = event.getMessage().getContentStripped();
-
-                //Get between all the spaces, to repmove the command bit
-                String[] args = message.split(" ");
-
-                if (args.length < 2) {
-                    event.processed = true;
-                    Reply(Settings.messagePrefix + "You forgot to add the message to make big!\nType `>help bigtext` for more information", event);
-                    return;
-                }
-
-                message = "";
-
-                for (int i = 1; i < args.length; i++) {
-                    message += args[i] + " ";
-                }
-
-                float fontSize = 800 - (message.length() * 10);
-
-                if (fontSize < 200) {
-                    fontSize = 200;
-                }
-
-                final BufferedImage image = new BufferedImage(4560, 1500,
-                        BufferedImage.TRANSLUCENT);
-
-
-                Graphics g = image.getGraphics();
-                g.setColor(Color.WHITE);
-                g.setFont(g.getFont().deriveFont(fontSize));
-                g.drawString(message, 10, 800);
-                g.dispose();
-
-
-                ByteArrayOutputStream baos = new ByteArrayOutputStream();
-                byte[] imageInByte;
-
-                try {
-                    ImageIO.write( image, "png", baos );
-
-                    baos.flush();
-                    imageInByte = baos.toByteArray();
-                    baos.close();
-
-                } catch (IOException e) {
-                    event.processed = true;
-                    Reply(Settings.messagePrefix+ "I'd appreciate it if you'd kindly stop trying to f*** up my program, k thx.", event);
-                    return;
-                }
-
-
-                ReplyImage(new ByteArrayInputStream(imageInByte), event);
-
+            if (event.getArgs().length < 2) {
                 event.processed = true;
+                Reply(Settings.messagePrefix + "You forgot to add the message to make big!\nType `>help bigtext` for more information", event);
+                return;
             }
+
+            String message = "";
+
+            for (int i = 1; i < event.getArgs().length; i++) {
+                message += event.getArgs()[i] + " ";
+            }
+
+            float fontSize = 800 - (message.length() * 10);
+
+            if (fontSize < 200) {
+                fontSize = 200;
+            }
+
+            final BufferedImage image = new BufferedImage(4560, 1500,
+                    BufferedImage.TRANSLUCENT);
+
+
+            Graphics g = image.getGraphics();
+            g.setColor(Color.WHITE);
+            g.setFont(g.getFont().deriveFont(fontSize));
+            g.drawString(message, 10, 800);
+            g.dispose();
+
+
+            ByteArrayOutputStream baos = new ByteArrayOutputStream();
+            byte[] imageInByte;
+
+            try {
+                ImageIO.write( image, "png", baos );
+
+                baos.flush();
+                imageInByte = baos.toByteArray();
+                baos.close();
+
+            } catch (IOException e) {
+                event.processed = true;
+                Reply(Settings.messagePrefix+ "I'd appreciate it if you'd kindly stop trying to f*** up my program, k thx.", event);
+                return;
+            }
+
+
+            ReplyImage(new ByteArrayInputStream(imageInByte), event);
+
+            event.processed = true;
         }).start();
     }
 }
